+++ components/ADempiere/TabManager/convenienceButtons/index.vue
@@ -23,99 +23,25 @@
       :container-uuid="tabAttributes.uuid"
     />
 
-    <el-button
-      v-if="isUndoChanges"
-      plain
-      size="small"
-      type="warning"
-      class="undo-changes-button"
-      @click="undoChanges()"
-    >
-      <svg-icon icon-class="undo" />
-      <span v-if="!isMobile">
-        {{ $t('actionMenu.undo') }}
-      </span>
-    </el-button>
+    <undo-change-button
+      :parent-uuid="parentUuid"
+      :container-uuid="tabAttributes.uuid"
+    />
 
-    <el-button
-      v-show="isRefreshRecord"
-      plain
-      size="small"
-      type="primary"
-      class="undo-changes-button"
-      @click="refreshCurrentRecord()"
-    >
-      <svg-icon icon-class="refresh" />
-      <span v-if="!isMobile">
-        {{ $t('actionMenu.refresh') }}
-      </span>
-    </el-button>
+    <refresh-record-button
+      :parent-uuid="parentUuid"
+      :container-uuid="tabAttributes.uuid"
+    />
 
-    <el-popover
-      v-if="isDeleteRecord"
-      v-model="isVisibleConfirmDelete"
-      placement="top"
-      class="delete-record-container"
-    >
-      <el-descriptions :title="$t('window.confirmDeleteRecord')" direction="vertical" :column="tabAttributes.identifierColumns.length" border>
-        <el-descriptions-item
-          v-for="(item, index) in tabAttributes.identifierColumns"
-          :key="index"
-          :label="item.name"
-        >
-          <cell-display-info
-            v-for="(record, key) in listOfRecordsToDeleted"
-            :key="key"
-            :field-attributes="item"
-            :data-row="record"
-          />
-        </el-descriptions-item>
-      </el-descriptions>
+    <delete-record-button
+      :parent-uuid="parentUuid"
+      :container-uuid="tabAttributes.uuid"
+    />
 
-      <div style="text-align: right; margin: 0">
-        <el-button size="mini" type="text" @click="isVisibleConfirmDelete = false">
-          {{ $t('window.cancel') }}
-        </el-button>
-        <el-button
-          ref="buttonConfirmDelete"
-          type="primary"
-          size="mini"
-          @click="deleteCurrentRecord()"
-        >
-          {{ $t('window.confirm') }}
-        </el-button>
-      </div>
-
-      <el-button
-        slot="reference"
-        plain
-        size="small"
-        type="danger"
-        class="delete-record-button"
-        @click="focusConfirmDelete()"
-      >
-        <svg-icon icon-class="delete" />
-        <span v-if="!isMobile">
-          {{ $t('actionMenu.delete') }}
-        </span>
-      </el-button>
-    </el-popover>
-
-    <el-button
-      v-show="isSaveRecord"
-      plain
-      size="small"
-      type="primary"
-      :loading="isSaveRecordLoading"
-      :disabled="isSaveRecordLoading"
-      class="undo-changes-button"
-      @click="saveChanges()"
-    >
-      <svg-icon icon-class="save-AD" />
-      <span v-if="!isMobile">
-        {{ $t('actionMenu.save') }}
-      </span>
-    </el-button>
+    <save-record-button
+      :parent-uuid="parentUuid"
+      :container-uuid="tabAttributes.uuid"
+    />
 
     <document-action
       v-if="tabAttributes.isDocument"
@@ -135,39 +61,30 @@
 </template>
 
 <script>
-import Vue from 'vue'
-import { defineComponent, computed, ref } from '@vue/composition-api'
+import { defineComponent, computed } from '@vue/composition-api'
 
 import store from '@/store'
-import router from '@/router'
-import language from '@/lang'
 
 // Components and Mixins
 import AdvancedTabQuery from '@theme/components/ADempiere/TabManager/AdvancedTabQuery.vue'
-import CellDisplayInfo from '@theme/components/ADempiere/DataTable/Components/CellDisplayInfo.vue'
 import DocumentAction from '@theme/components/ADempiere/TabManager/convenienceButtons/documentAction.vue'
 import NewRecordButton from '@theme/components/ADempiere/TabManager/convenienceButtons/NewRecordButton.vue'
-
-// Constants
-import { LOG_COLUMNS_NAME_LIST } from '@/utils/ADempiere/constants/systemColumns'
-
-// Utils and Melper Methods
-import { showMessage } from '@/utils/ADempiere/notification'
-import { isEmptyValue } from '@/utils/ADempiere/valueUtils'
-import {
-  refreshRecord,
-  deleteRecord,
-  undoChange
-} from '@/utils/ADempiere/dictionary/window'
+import SaveRecordButton from '@theme/components/ADempiere/TabManager/convenienceButtons/SaveRecordButton.vue'
+import UndoChangeButton from '@theme/components/ADempiere/TabManager/convenienceButtons/UndoChangeButton.vue'
+import RefreshRecordButton from '@theme/components/ADempiere/TabManager/convenienceButtons/RefreshRecordButton.vue'
+import DeleteRecordButton from '@theme/components/ADempiere/TabManager/convenienceButtons/DeleteRecordButton.vue'
 
 export default defineComponent({
   name: 'ConvenienceButtons',
 
   components: {
     AdvancedTabQuery,
-    CellDisplayInfo,
     DocumentAction,
-    NewRecordButton
+    NewRecordButton,
+    SaveRecordButton,
+    UndoChangeButton,
+    RefreshRecordButton,
+    DeleteRecordButton
   },
 
   props: {
@@ -191,31 +108,9 @@ export default defineComponent({
 
   setup(props) {
     const containerUuid = props.tabAttributes.uuid
-    const isVisibleConfirmDelete = ref(false)
-    const buttonConfirmDelete = ref(null)
-    const isSaveRecordLoading = ref(false)
-
-    const currentRoute = router.app._route
 
     const recordUuid = computed(() => {
       return store.getters.getUuidOfContainer(containerUuid)
-    })
-
-    const isMobile = computed(() => {
-      return store.state.app.device === 'mobile'
-    })
-
-    const listOfRecordsToDeleted = computed(() => {
-      if (!getCurrentTab.value.isShowedTableRecords) {
-        const record = store.getters.getTabCurrentRow({
-          containerUuid
-        })
-        if (isEmptyValue(record)) {
-          return []
-        }
-        return [record]
-      }
-      return selectionsRecords.value
     })
 
     const isDisableOptionsTabChild = computed(() => {
@@ -228,71 +123,6 @@ export default defineComponent({
       return true
     })
 
-    const selectionsRecords = computed(() => {
-      return props.containerManager.getSelection({
-        containerUuid
-      })
-    })
-
-    const isExistsChanges = computed(() => {
-      const persistenceValues = store.getters.getPersistenceAttributesChanges({
-        parentUuid: props.parentUuid,
-        containerUuid,
-        recordUuid: recordUuid.value
-      })
-      return !isEmptyValue(persistenceValues)
-    })
-
-    const isRefreshRecord = computed(() => {
-      if (isEmptyValue(recordUuid.value)) {
-        return false
-      }
-      return !isExistsChanges.value
-    })
-
-    const emptyMandatoryFields = computed(() => {
-      return store.getters.getTabFieldsEmptyMandatory({
-        parentUuid: props.parentUuid,
-        containerUuid,
-        formatReturn: false
-      }).filter(itemField => {
-        // omit send to server (to create or update) columns manage by backend
-        return itemField.isAlwaysUpdateable ||
-          !LOG_COLUMNS_NAME_LIST.includes(itemField.columnName)
-      }).map(itemField => {
-        return itemField.name
-      })
-    })
-
-    const isSaveRecord = computed(() => {
-      if (!isEmptyValue(emptyMandatoryFields.value)) {
-        return false
-      }
-
-      return isExistsChanges.value
-    })
-
-    const isDeleteRecord = computed(() => {
-      if (isExistsChanges.value) {
-        return false
-      }
-      return deleteRecord.enabled({
-        parentUuid: props.parentUuid,
-        tabParentIndex: props.tabAttributes.tabParentIndex,
-        containerUuid
-      })
-    })
-
-    const isUndoChanges = computed(() => {
-      if (!isEmptyValue(recordUuid.value)) {
-        return isExistsChanges.value
-      }
-
-      // without old record
-      const oldRecordUuid = store.getters.getCurrentRecordOnPanel(containerUuid)
-      return !isEmptyValue(oldRecordUuid) || isExistsChanges.value
-    })
-
     const getCurrentTab = computed(() => {
       const tab = store.getters.getStoredTab(
         props.parentUuid,
@@ -303,149 +133,6 @@ export default defineComponent({
       }
       return props.tabAttributes
     })
-
-    function refreshCurrentRecord() {
-      refreshRecord.refreshRecord({
-        parentUuid: props.parentUuid,
-        containerUuid
-      })
-
-      const info = {
-        fieldsList: props.containerManager.getFieldsList({
-          parentUuid: props.parentUuid,
-          containerUuid: containerUuid
-        }),
-        option: language.t('actionMenu.refresh')
-      }
-      store.dispatch('fieldListInfo', { info })
-    }
-
-    function focusConfirmDelete() {
-      if (buttonConfirmDelete.value) {
-        Vue.nextTick(() => {
-          // TODO: Doesn't work, focus confirm button with displayed popover.
-          // buttonConfirmDelete.value.$el.focus()
-        })
-      }
-    }
-
-    function deleteCurrentRecord() {
-      if (getCurrentTab.value.isShowedTableRecords && !isEmptyValue(selectionsRecords.value)) {
-        store.dispatch('deleteSelectedRecordsFromWindow', {
-          parentUuid: props.parentUuid,
-          containerUuid
-        })
-        isVisibleConfirmDelete.value = false
-        return
-      }
-
-      const info = {
-        fieldsList: props.tabAttributes.fieldsList,
-        option: language.t('actionMenu.delete')
-      }
-      store.dispatch('fieldListInfo', { info })
-      deleteRecord.deleteRecord({
-        parentUuid: props.parentUuid,
-        containerUuid,
-        recordUuid: recordUuid.value
-      })
-      isVisibleConfirmDelete.value = false
-    }
-
-    function undoChanges() {
-      // store.dispatch('fieldListInfo', {
-      //   fieldsList: props.tabAttributes.fieldsList,
-      //   option: language.t('actionMenu.undo')
-      // })
-      const info = {
-        fieldsList: props.containerManager.getFieldsList({
-          parentUuid: props.parentUuid,
-          containerUuid: containerUuid
-        }),
-        option: language.t('actionMenu.undo')
-      }
-
-      store.dispatch('fieldListInfo', { info })
-      undoChange.undoChange({
-        parentUuid: props.parentUuid,
-        containerUuid
-      })
-    }
-
-    function saveChanges() {
-      const emptyMandatory = emptyMandatoryFields.value
-
-      if (!isEmptyValue(emptyMandatory)) {
-        showMessage({
-          message: language.t('notifications.mandatoryFieldMissing') + emptyMandatory,
-          type: 'info'
-        })
-        return
-      }
-
-      const info = {
-        fieldsList: props.containerManager.getFieldsList({
-          parentUuid: props.parentUuid,
-          containerUuid: containerUuid
-        }),
-        option: language.t('actionMenu.save')
-      }
-
-      store.dispatch('fieldListInfo', { info })
-      isSaveRecordLoading.value = true
-
-      store.dispatch('flushPersistenceQueue', {
-        parentUuid: props.parentUuid,
-        containerUuid,
-        tableName: props.tabAttributes.tableName,
-        recordUuid: recordUuid.value
-      })
-        .then(response => {
-          const {
-            name,
-            query,
-            params
-          } = currentRoute
-          const { id } = response
-          // refresh parent tab on document window
-          if (!props.tabAttributes.isParentTab) {
-            const { firstTabUuid } = props.tabAttributes
-            const firstTab = store.getters.getStoredTab(
-              props.parentUuid,
-              firstTabUuid
-            )
-            if (!isEmptyValue(firstTab) && firstTab.isDocument) {
-              refreshRecord.refreshRecord({
-                parentUuid: props.parentUuid,
-                containerUuid: firstTabUuid
-              })
-            }
-          }
-
-          router.replace({
-            name,
-            query: {
-              ...query,
-              recordId: id,
-              filters: []
-            },
-            params: {
-              ...params,
-              filters: []
-            }
-          }, () => {})
-        })
-        .catch(error => {
-          console.error('Error saving record', error.message)
-          showMessage({
-            message: error.message,
-            type: 'error'
-          })
-        })
-        .finally(() => {
-          isSaveRecordLoading.value = false
-        })
-    }
 
     function openLog() {
       const list = store.getters.getTabRecordsList({ containerUuid })
@@ -479,27 +166,11 @@ export default defineComponent({
     // })
 
     return {
-      buttonConfirmDelete,
-      isVisibleConfirmDelete,
-      isSaveRecordLoading,
       // computed
-      isMobile,
       recordUuid,
-      selectionsRecords,
-      isExistsChanges,
-      isDeleteRecord,
-      isRefreshRecord,
-      isUndoChanges,
       getCurrentTab,
       isDisableOptionsTabChild,
-      isSaveRecord,
-      listOfRecordsToDeleted,
       // methods
-      deleteCurrentRecord,
-      focusConfirmDelete,
-      refreshCurrentRecord,
-      undoChanges,
-      saveChanges,
       openLog
     }
   }
